fix: edit according comments
