+++ package.json
@@ -10,7 +10,7 @@
     "focus-trap-react": "^10.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.10.0",
+    "react-router-dom": "6.10.0",
     "react-scripts": "5.0.1",
     "sass": "^1.59.2",
     "web-vitals": "^2.1.4"
