+++ src/components/footer/Footer.js
@@ -2,12 +2,10 @@ import React from "react";
 import './Footer.scss';
 import {Logo} from "../index";
 
-const Footer = () => {
-    return (
-        <footer>
-            <Logo/>
-        </footer>
-    )
-}
+const Footer = () => (
+    <footer>
+        <Logo/>
+    </footer>
+)
 
 export default Footer;
