+++ src/components/logo/Logo.js
@@ -1,10 +1,6 @@
 import React from "react";
 import './Logo.scss';
 
-const Logo = () => {
-    return (
-        <div className='logo'><span>netflix</span>roulette</div>
-    )
-}
+const Logo = () => (<div className='logo'><span>netflix</span>roulette</div>)
 
 export default Logo;
