+++ src/components/movieDetails/movieDetails.test.js
@@ -2,12 +2,11 @@ import React from "react";
 import {MovieDetails} from "../../components";
 import { render } from "@testing-library/react";
 import {moviesArray} from "../../mockedMovies";
+import {MemoryRouter} from "react-router-dom";
 
 describe("MovieDetails", () => {
     it("renders snapshot", () => {
-        expect.assertions(1);
-
-        const { asFragment } = render(<MovieDetails movie={moviesArray[0]}/>);
+        const { asFragment } = render(<MemoryRouter><MovieDetails movie={moviesArray[0]}/></MemoryRouter>);
 
         expect(asFragment()).toMatchSnapshot();
     })
