+++ src/constant.js
@@ -1,0 +1,1 @@
+export const BASE_URL = "http://localhost:4000";
