+++ src/utils.js
@@ -1,3 +1,18 @@
 const convertDateIntoYear = (year) => new Date(year).getFullYear();
 
-export default convertDateIntoYear;
+const fetchData = async (url, callback) => {
+    let controller, signal;
+    if (controller) {
+        controller.abort()
+    }
+    controller = new AbortController();
+    signal = controller.signal;
+
+    const response = await fetch(url, {
+        signal: signal,
+    })
+    const data = await response.json();
+    controller = null;
+    callback(data);
+}
+export {convertDateIntoYear, fetchData};
